requirements.txt
aiohappyeyeballs
aiohttp
aiohttp-retry
aiosignal
annotated-types
anyio
async-timeout
attrs
blinker
certifi
charset-normalizer
click
dataclasses-json
distro
exceptiongroup
faiss-cpu
filelock
Flask
frozenlist
fsspec
h11
httpcore
httpx
huggingface-hub
idna
importlib_metadata
itsdangerous
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jupyter
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langchain-together
langsmith
MarkupSafe
marshmallow
mpmath
multidict
mypy-extensions
networkx
numpy
openai
orjson
packaging
pillow
pydantic
python-dotenv
pydantic_core
PyJWT
pytube
PyYAML
regex
requests
safetensors
scikit-learn
scipy
sentence-transformers
sniffio
SQLAlchemy
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
torch
tqdm
transformers
twilio
typing-inspect
typing_extensions
tzdata
urllib3
Werkzeug
yarl
youtube-transcript-api
zipp
